move swagger libs to /src/libs
