+++ src/app/(main)/api-doc/page.tsx
@@ -1,7 +1,7 @@
 import "swagger-ui-react/swagger-ui.css";
 
-import { getApiDocs } from "../../lib/swagger";
-import ReactSwagger from "./react-swagger";
+import { getApiDocs } from "@/lib/api/swagger";
+import ReactSwagger from "@/lib/api/react-swagger";
 import { Metadata } from "next";
 
 export const metadata: Metadata = {
